refactor(chat): chatRoomSlice 분리, 중복 유틸 제거, memoized selector 추가
- chatStoreUtils.js: toId/toIso/sortByLatest 등 공통 유틸 추출 (3곳 중복 제거)
- chatReadThunk.js: markReadThunk/fetchReadPointersThunk를 slice에서 분리
- chatRoomExtraReducers.js: extraReducers 빌더 함수 분리
- chatRoomSlice.js: 628줄 → 254줄, createSelector 기반 memoized selector 적용
- selectors.js: 전역 memoized selector 4개 추가
